Began validations

+++ routes/courses.js
@@ -3,18 +3,27 @@ const Course = require("../models/course");
 const auth = require("../middleware/auth");
 const router = Router();
 
+function isOwner(course, req) {
+  return course.userId.toString() === req.user._id.toString();
+}
+
 router.get("/", async (req, res) => {
-  const courses = await Course.find()
-    .populate("userId", "email name")
-    .select("price title img");
+  try {
+    const courses = await Course.find()
+      .populate("userId", "email name")
+      .select("price title img");
 
-  console.log(courses);
+    console.log(courses);
 
-  res.render("courses", {
-    title: "Курсы",
-    isCourses: true,
-    courses
-  });
+    res.render("courses", {
+      title: "Курсы",
+      isCourses: true,
+      userId: req.user ? req.user._id.toString() : null,
+      courses
+    });
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 router.get("/:id/edit", auth, async (req, res) => {
@@ -22,24 +31,42 @@ router.get("/:id/edit", auth, async (req, res) => {
     return res.redirect("/");
   }
 
-  const course = await Course.findById(req.params.id);
+  try {
+    const course = await Course.findById(req.params.id);
 
-  res.render("course-edit", {
-    title: `Редактировать ${course.title}`,
-    course
-  });
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+
+    res.render("course-edit", {
+      title: `Редактировать ${course.title}`,
+      course
+    });
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 router.post("/edit", auth, async (req, res) => {
-  const { id } = req.body;
-  delete req.body.id;
-  await Course.findByIdAndUpdate(id, req.body);
-  res.redirect("/courses");
+  try {
+    const { id } = req.body;
+    delete req.body.id;
+    const course = await Course.findById(id);
+
+    if (!isOwner(course, req)) {
+      return res.redirect("/courses");
+    }
+    Object.assign(course, req.body);
+    await course.save();
+    res.redirect("/courses");
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 router.post("/remove", auth, async (req, res) => {
   try {
-    await Course.deleteOne({ _id: req.body.id });
+    await Course.deleteOne({ _id: req.body.id, userId: req.user._id });
     res.redirect("/courses");
   } catch (err) {
     console.log(err);
@@ -47,13 +74,16 @@ router.post("/remove", auth, async (req, res) => {
 });
 
 router.get("/:id", async (req, res) => {
-  console.log(`ID ${req.params.id}`);
-  const course = await Course.findById(req.params.id);
-  res.render("course", {
-    layout: "empty",
-    title: `Курс ${course.title}`,
-    course
-  });
+  try {
+    const course = await Course.findById(req.params.id);
+    res.render("course", {
+      layout: "empty",
+      title: `Курс ${course.title}`,
+      course
+    });
+  } catch (err) {
+    console.log(err);
+  }
 });
 
 module.exports = router;
